requirements.txt
PyQt6
pandas
numpy
websockets
requests
aiohttp
aiosqlite
python-dotenv
keyring
cryptography
torch
qasync
ta-lib
PyYAML
pytest
pytest-asyncio
psutil
holidays
pyqtgraph
qtawesome
